@@ -13,18 +14,17 @@ if (keystorePropertiesFile.exists() && keystorePropertiesFile.canRead()) {
 }
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 30
+    buildToolsVersion '29.0.3'
     defaultConfig {
         applicationId "io.github.gmathi.novellibrary"
-        minSdkVersion 16
-        targetSdkVersion 27
+        minSdkVersion 21
+        targetSdkVersion 30
         multiDexEnabled true
-        versionCode 49
-        versionName "0.9.beta"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        versionCode 95
+        versionName "0.16.beta"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
-        resConfigs "auto"
     }
 
     signingConfigs {
