@@ -1,9 +1,25 @@
 package io.github.gmathi.novellibrary.activity
 
-import android.support.v7.app.AppCompatActivity
+import android.content.Context
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.google.firebase.analytics.FirebaseAnalytics
+import com.google.firebase.analytics.ktx.analytics
+import com.google.firebase.ktx.Firebase
+import io.github.gmathi.novellibrary.util.lang.LocaleManager
 
 
 abstract class BaseActivity : AppCompatActivity() {
-    //Empty Base Activity
 
+    lateinit var firebaseAnalytics: FirebaseAnalytics
+
+    override fun attachBaseContext(newBase: Context) {
+        super.attachBaseContext(LocaleManager.updateContextLocale(newBase))
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        // Obtain the FirebaseAnalytics instance.
+        firebaseAnalytics = Firebase.analytics
+    }
 }
