@@ -1,32 +1,55 @@
 package io.github.gmathi.novellibrary.adapter
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
+import io.github.gmathi.novellibrary.R
 import io.github.gmathi.novellibrary.util.inflate
 import java.util.*
+import java.util.concurrent.atomic.AtomicBoolean
+import kotlin.collections.ArrayList
 
-class GenericAdapter<T>(val items: ArrayList<T>, val layoutResId: Int, val listener: Listener<T>) : RecyclerView.Adapter<GenericAdapter.ViewHolder<T>>() {
+class GenericAdapter<T>(val items: ArrayList<T>, val layoutResId: Int, val listener: Listener<T>, var loadMoreListener: LoadMoreListener? = null) : RecyclerView.Adapter<GenericAdapter.ViewHolder<T>>() {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) = ViewHolder<T>(parent.inflate(layoutResId))
+    companion object {
+        const val VIEW_TYPE_NORMAL = 0
+        const val VIEW_TYPE_LOAD_MORE = 1
+    }
 
-    override fun onBindViewHolder(holder: ViewHolder<T>, position: Int) = holder.bind(item = items[position], listener = listener, position = position)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) = if (viewType == VIEW_TYPE_NORMAL) ViewHolder<T>(parent.inflate(layoutResId)) else ViewHolder(parent.inflate(R.layout.listitem_progress_bar))
 
-    override fun onBindViewHolder(holder: ViewHolder<T>, position: Int, payloads: MutableList<Any>) = holder.bind(item = items[position], listener = listener, position = position, payloads = payloads)
+    override fun onBindViewHolder(holder: ViewHolder<T>, position: Int) = if (position == items.size) holder.loadMore(loadMoreListener) else holder.bind(item = items[position], listener = listener, position = position)
 
-    override fun getItemCount() = items.size
+    override fun onBindViewHolder(holder: ViewHolder<T>, position: Int, payloads: MutableList<Any>) {
+        if (items.size <= position + (loadMoreListener?.preloadCount ?: 0))
+            holder.loadMore(loadMoreListener)
+        if (position != items.size)
+            holder.bind(item = items[position], listener = listener, position = position, payloads = payloads)
+    }
+
+    override fun getItemCount() = if (loadMoreListener != null) items.size + 1 else items.size
+
+    override fun getItemViewType(position: Int): Int {
+        return if (loadMoreListener != null && position == items.size)
+            VIEW_TYPE_LOAD_MORE
+        else
+            VIEW_TYPE_NORMAL
+    }
 
     class ViewHolder<T>(itemView: View) : RecyclerView.ViewHolder(itemView) {
         fun bind(item: T, listener: Listener<T>, position: Int) {
-            with(itemView) { setOnClickListener { listener.onItemClick(item) } }
+            with(itemView) { setOnClickListener { listener.onItemClick(item, position) } }
             listener.bind(item = item, itemView = itemView, position = position)
         }
 
         fun bind(item: T, listener: Listener<T>, position: Int, payloads: MutableList<Any>?) {
-            with(itemView) { setOnClickListener { listener.onItemClick(item) } }
+            with(itemView) { setOnClickListener { listener.onItemClick(item, position) } }
             listener.bind(item = item, itemView = itemView, position = position, payloads = payloads)
         }
 
+        fun loadMore(loadMoreListener: LoadMoreListener?) {
+            loadMoreListener?.loadMore()
+        }
     }
 
     interface Listener<in T> {
@@ -34,8 +57,14 @@ class GenericAdapter<T>(val items: ArrayList<T>, val layoutResId: Int, val liste
         fun bind(item: T, itemView: View, position: Int, payloads: MutableList<Any>?) {
             bind(item, itemView, position)
         }
+        fun onItemClick(item: T, position: Int)
+    }
 
-        fun onItemClick(item: T)
+    interface LoadMoreListener {
+        var currentPageNumber: Int
+        val preloadCount: Int
+        val isPageLoading: AtomicBoolean
+        fun loadMore()
     }
 
     fun updateData(newItems: ArrayList<T>) {
@@ -57,6 +86,15 @@ class GenericAdapter<T>(val items: ArrayList<T>, val layoutResId: Int, val liste
         items.clear()
         items.addAll(newItems)
         notifyDataSetChanged()
+    }
+
+    fun addItems(newItems: ArrayList<T>) {
+        items.addAll(newItems)
+        notifyDataSetChanged()
+    }
+
+    fun addItems(newItems: List<T>) {
+        addItems(ArrayList(newItems))
     }
 
     fun updateItem(item: T) {
