@@ -1,21 +1,21 @@
 package io.github.gmathi.novellibrary.adapter
 
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentStatePagerAdapter
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentStatePagerAdapter
 import java.util.*
 
 
-class GenericFragmentStatePagerAdapter(manager: FragmentManager, val titles: Array<String>?, val pagerCount: Int, val listener: Listener) : FragmentStatePagerAdapter(manager) {
+class GenericFragmentStatePagerAdapter(manager: FragmentManager, private val titles: Array<String>?, private val pagerCount: Int, val listener: Listener) : FragmentStatePagerAdapter(manager) {
 
-    override fun getItem(position: Int): Fragment? = listener.getFragmentForItem(position)
+    override fun getItem(position: Int): Fragment = listener.getFragmentForItem(position)
 
     override fun getCount(): Int = pagerCount
 
     override fun getPageTitle(position: Int): CharSequence? = if (titles != null) titles[position].toUpperCase(Locale.getDefault()) else super.getPageTitle(position)
 
     interface Listener {
-        fun getFragmentForItem(position: Int): Fragment?
+        fun getFragmentForItem(position: Int): Fragment
     }
 
 }
