@@ -1,32 +1,35 @@
 package io.github.gmathi.novellibrary
 
-import android.annotation.SuppressLint
-import android.app.NotificationChannel
-import android.app.NotificationManager
 import android.content.Context
-import android.os.Build
-import android.support.multidex.MultiDex
-import android.support.multidex.MultiDexApplication
-import android.support.v7.app.AppCompatDelegate
-import android.util.Log
+import android.content.res.Configuration
 import android.webkit.WebView
-import com.squareup.leakcanary.LeakCanary
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.core.app.NotificationManagerCompat
+import androidx.multidex.MultiDexApplication
+import com.google.android.gms.common.GooglePlayServicesNotAvailableException
+import com.google.android.gms.security.ProviderInstaller
+import com.google.firebase.remoteconfig.FirebaseRemoteConfig
+import com.google.firebase.remoteconfig.FirebaseRemoteConfigSettings
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 import io.github.gmathi.novellibrary.database.DBHelper
+import io.github.gmathi.novellibrary.database.deleteWebPageSettings
+import io.github.gmathi.novellibrary.database.deleteWebPages
+import io.github.gmathi.novellibrary.model.other.SelectorQuery
 import io.github.gmathi.novellibrary.network.HostNames
+import io.github.gmathi.novellibrary.network.MultiTrustManager
 import io.github.gmathi.novellibrary.service.sync.BackgroundNovelSyncTask
+import io.github.gmathi.novellibrary.util.Constants
 import io.github.gmathi.novellibrary.util.DataCenter
-import io.github.gmathi.novellibrary.util.Utils
+import io.github.gmathi.novellibrary.util.Logs
+import io.github.gmathi.novellibrary.util.lang.LocaleManager
 import java.io.File
 import java.security.KeyManagementException
 import java.security.NoSuchAlgorithmException
 import java.security.SecureRandom
-import java.security.cert.CertificateException
-import java.security.cert.X509Certificate
+import java.util.*
 import javax.net.ssl.HttpsURLConnection
 import javax.net.ssl.SSLContext
-import javax.net.ssl.X509TrustManager
-import com.google.android.gms.common.GooglePlayServicesNotAvailableException
-import com.google.android.gms.security.ProviderInstaller
 
 
 val dataCenter: DataCenter by lazy {
@@ -43,26 +46,45 @@ class NovelLibraryApplication : MultiDexApplication() {
         var dbHelper: DBHelper? = null
 
         private const val TAG = "NovelLibraryApplication"
+
+        fun refreshDBHelper(context: Context) {
+            dbHelper = DBHelper.refreshInstance(context)
+        }
     }
 
     override fun onCreate() {
         dataCenter = DataCenter(applicationContext)
         dbHelper = DBHelper.getInstance(applicationContext)
-
+        val date = Calendar.getInstance()
+        if (date.get(Calendar.MONTH) == 4 && date.get(Calendar.DAY_OF_MONTH) == 1) {
+            if (!dataCenter?.fooled!!) {
+                dataCenter?.language = "pa"
+                dataCenter?.fooled = true
+            }
+        } else dataCenter?.fooled = false
         super.onCreate()
 
+        if (dataCenter?.hasAlreadyDeletedOldChannels == false) {
+            deleteOldNotificationChannels()
+        }
+
+        //Stray webPages to be deleted
+        dbHelper?.deleteWebPages(-1L)
+        dbHelper?.deleteWebPageSettings(-1L)
+
+//        dataCenter?.isDeveloper = true
+//        dataCenter?.lockRoyalRoad = false
+//
         try {
             HostNames.hostNamesList = dataCenter!!.getVerifiedHosts()
+            HostNames.defaultHostNamesList.forEach {
+                HostNames.addHost(it)
+            }
         } catch (e: Exception) {
-            Utils.error(TAG, "Set the HostNames.hostNamesList from dataCenter", e)
+            Logs.error(TAG, "Set the HostNames.hostNamesList from dataCenter", e)
         }
 
         AppCompatDelegate.setCompatVectorFromResourcesEnabled(true)
-
-        if (LeakCanary.isInAnalyzerProcess(this)) {
-            return
-        }
-        LeakCanary.install(this)
 
         val imagesDir = File(filesDir, "images")
         if (!imagesDir.exists())
@@ -71,44 +93,40 @@ class NovelLibraryApplication : MultiDexApplication() {
         try {
             enableSSLSocket()
         } catch (e: Exception) {
-            Log.e(TAG, "enableSSLSocket(): ${e.localizedMessage}", e)
+            Logs.error(TAG, "enableSSLSocket(): ${e.localizedMessage}", e)
         }
 
         //BugFix for <5.0 devices
         //https://stackoverflow.com/questions/29916962/javax-net-ssl-sslhandshakeexception-javax-net-ssl-sslprotocolexception-ssl-han
         updateAndroidSecurityProvider()
 
-        if (BuildConfig.DEBUG && Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+        if (BuildConfig.DEBUG) {
             WebView.setWebContentsDebuggingEnabled(true)
         }
 
-        initChannels(applicationContext)
-
         if (dataCenter!!.enableNotifications)
             startSyncService()
+
+        setRemoteConfig()
+    }
+
+    @Deprecated("This method deletes old notification channels. Assuming that all users updated and run the app at least once, this method should be removed!")
+    private fun deleteOldNotificationChannels() {
+        val notificationManager = NotificationManagerCompat.from(applicationContext)
+        notificationManager.deleteNotificationChannel("default")
+        notificationManager.deleteNotificationChannel("io.github.gmathi.novellibrary.service.tts.NOW_PLAYING")
+        dataCenter?.hasAlreadyDeletedOldChannels = true
     }
 
     @Throws(KeyManagementException::class, NoSuchAlgorithmException::class)
     private fun enableSSLSocket() {
-
         HttpsURLConnection.setDefaultHostnameVerifier { hostName: String?, _ ->
             if (hostName != null) HostNames.isVerifiedHost(hostName) else false
         }
-
         val context = SSLContext.getInstance("TLS")
-        context.init(null, arrayOf<X509TrustManager>(object : X509TrustManager {
-            @SuppressLint("TrustAllX509TrustManager")
-            @Throws(CertificateException::class)
-            override fun checkClientTrusted(chain: Array<X509Certificate>, authType: String) {
-            }
-
-            @SuppressLint("TrustAllX509TrustManager")
-            @Throws(CertificateException::class)
-            override fun checkServerTrusted(chain: Array<X509Certificate>, authType: String) {
-            }
-
-            override fun getAcceptedIssuers(): Array<X509Certificate?> = arrayOfNulls(0)
-        }), SecureRandom())
+        val multiTrustManager = MultiTrustManager()
+        multiTrustManager.addDefaultTrustManager()
+        context.init(null, arrayOf(multiTrustManager), SecureRandom())
         HttpsURLConnection.setDefaultSSLSocketFactory(context.socketFactory)
     }
 
@@ -116,22 +134,11 @@ class NovelLibraryApplication : MultiDexApplication() {
         try {
             ProviderInstaller.installIfNeeded(this)
         } catch (e: GooglePlayServicesNotAvailableException) {
-            Utils.error("SecurityException", "Google Play Services not available.")
+            Logs.error("SecurityException", "Google Play Services not available.")
         } catch (e: Exception) {
-            Utils.error("Exception", "Other Exception: ${e.localizedMessage}", e)
-        }
-    }
+            Logs.error("Exception", "Other Exception: ${e.localizedMessage}", e)
 
-    private fun initChannels(context: Context) {
-        if (Build.VERSION.SDK_INT < 26) {
-            return
         }
-        val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        val channel = NotificationChannel(getString(R.string.default_notification_channel_id),
-                "Default Channel",
-                NotificationManager.IMPORTANCE_DEFAULT)
-        channel.description = "Default Channel Description"
-        notificationManager.createNotificationChannel(channel)
     }
 
     private fun startSyncService() {
@@ -139,9 +146,23 @@ class NovelLibraryApplication : MultiDexApplication() {
     }
 
     override fun attachBaseContext(base: Context) {
-        super.attachBaseContext(base)
-        MultiDex.install(this)
+        super.attachBaseContext(LocaleManager.updateContextLocale(base))
     }
 
+    override fun onConfigurationChanged(newConfig: Configuration) {
+        super.onConfigurationChanged(newConfig)
+        LocaleManager.updateContextLocale(this)
+    }
+
+    fun setRemoteConfig() {
+        val remoteConfig = FirebaseRemoteConfig.getInstance()
+        remoteConfig.setConfigSettingsAsync(FirebaseRemoteConfigSettings.Builder().build())
+        val defaults = HashMap<String, Any>()
+        defaults[Constants.RemoteConfig.SELECTOR_QUERIES] = "[]"
+        remoteConfig.setDefaultsAsync(defaults)
+        remoteConfig.fetchAndActivate().addOnCompleteListener {
+            dataCenter?.htmlCleanerSelectorQueries = Gson().fromJson(remoteConfig.getString(Constants.RemoteConfig.SELECTOR_QUERIES), object : TypeToken<ArrayList<SelectorQuery>>() {}.type)
+        }
+    }
 
 }
