@@ -1,41 +1,56 @@
 package io.github.gmathi.novellibrary.activity
 
 import android.annotation.SuppressLint
+import android.content.ClipData
+import android.content.ClipboardManager
+import android.content.Context
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
 import android.text.Html
 import android.text.Spannable
 import android.text.SpannableString
 import android.text.TextPaint
 import android.text.method.LinkMovementMethod
 import android.text.style.ClickableSpan
-import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
 import android.widget.LinearLayout
 import android.widget.TextView
-import co.metalab.asyncawait.async
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.lifecycleScope
 import com.afollestad.materialdialogs.MaterialDialog
 import com.afollestad.materialdialogs.Theme
 import com.bumptech.glide.Glide
+import com.google.firebase.analytics.ktx.logEvent
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 import io.github.gmathi.novellibrary.R
 import io.github.gmathi.novellibrary.dataCenter
-import io.github.gmathi.novellibrary.database.getNovel
-import io.github.gmathi.novellibrary.database.insertNovel
-import io.github.gmathi.novellibrary.database.updateNovel
+import io.github.gmathi.novellibrary.database.*
 import io.github.gmathi.novellibrary.dbHelper
-import io.github.gmathi.novellibrary.model.Novel
+import io.github.gmathi.novellibrary.extensions.*
+import io.github.gmathi.novellibrary.model.database.Novel
 import io.github.gmathi.novellibrary.network.NovelApi
 import io.github.gmathi.novellibrary.network.getNovelDetails
+import io.github.gmathi.novellibrary.network.sync.NovelSync
 import io.github.gmathi.novellibrary.util.*
+import io.github.gmathi.novellibrary.util.system.*
+import io.github.gmathi.novellibrary.util.view.*
 import kotlinx.android.synthetic.main.activity_novel_details.*
 import kotlinx.android.synthetic.main.content_novel_details.*
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import kotlin.math.min
 
 
 class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener {
+
+    companion object {
+        const val TAG = "NovelDetailsActivity"
+    }
 
     lateinit var novel: Novel
 
@@ -75,30 +90,36 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
         if (!Utils.isConnectedToNetwork(this)) {
             if (novel.id == -1L) {
                 swipeRefreshLayout.isRefreshing = false
-                progressLayout.showError(ContextCompat.getDrawable(this, R.drawable.ic_warning_white_vector), getString(R.string.no_internet), getString(R.string.try_again), {
+                progressLayout.noInternetError {
                     progressLayout.showLoading()
                     getNovelInfo()
-                })
+                }
             }
             return
         }
 
-        async {
+        lifecycleScope.launch {
             try {
-                val downloadedNovel = await { NovelApi.getNovelDetails(novel.url) }
+                val downloadedNovel = withContext(Dispatchers.IO) { NovelApi.getNovelDetails(novel.url) }
                 novel.copyFrom(downloadedNovel)
                 addNovelToHistory()
-                if (novel.id != -1L) await { dbHelper.updateNovel(novel) }
+                if (novel.id != -1L) withContext(Dispatchers.IO) { dbHelper.updateNovel(novel) }
                 setupViews()
                 progressLayout.showContent()
                 swipeRefreshLayout.isRefreshing = false
             } catch (e: Exception) {
                 e.printStackTrace()
+                val errorMessage = e.localizedMessage ?: "Unknown Error" + "\n" + e.stackTrace.joinToString(separator = "\n") { it.toString() }
+                val clipboard = getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+                val clip: ClipData = ClipData.newPlainText("Error Message", errorMessage)
+                clipboard.setPrimaryClip(clip)
+                MaterialDialog.Builder(this@NovelDetailsActivity).title("Error!").content("The error message has been copied to clipboard. Please paste it and send it the developer in discord.")
+                    .show()
                 if (novel.id == -1L)
-                    progressLayout.showError(ContextCompat.getDrawable(this@NovelDetailsActivity, R.drawable.ic_warning_white_vector), getString(R.string.failed_to_load_url), getString(R.string.try_again), {
+                    progressLayout.showError(errorText = getString(R.string.failed_to_load_url), buttonText = getString(R.string.try_again)) {
                         progressLayout.showLoading()
                         getNovelInfo()
-                    })
+                    }
             }
         }
     }
@@ -106,17 +127,31 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
     @SuppressLint("SetTextI18n")
     private fun setupViews() {
         setNovelImage()
+
         novelDetailsName.applyFont(assets).text = novel.name
+        novelDetailsName.isSelected = dataCenter.enableScrollingText
+
+        val listener: View.OnClickListener = View.OnClickListener {
+            MaterialDialog.Builder(this)
+                .title("Novel Name")
+                .content(novel.name)
+                .build().show()
+        }
+        novelDetailsName.setOnClickListener(listener)
+        novelDetailsNameInfo.setOnClickListener(listener)
+
         setNovelAuthor()
 
         novelDetailsStatus.applyFont(assets).text = "N/A"
-        if (novel.metaData["Year"] != null)
-            novelDetailsStatus.applyFont(assets).text = novel.metaData["Year"]
+        if (novel.metadata["Year"] != null)
+            novelDetailsStatus.applyFont(assets).text = novel.metadata["Year"]
 
+        setLicensingInfo()
         setNovelRating()
         setNovelAddToLibraryButton()
         setNovelGenre()
         setNovelDescription()
+
         novelDetailsChapters.text = getString(R.string.chapters) + " (${novel.chaptersCount})"
         novelDetailsChaptersLayout.setOnClickListener {
             if (novel.chaptersCount != 0L) startChaptersActivity(novel, false)
@@ -126,21 +161,35 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
     }
 
     private fun setNovelImage() {
-        if (novel.imageUrl != null) {
+        if (!novel.imageUrl.isNullOrBlank()) {
             Glide.with(this)
-                    .load(novel.imageUrl?.getGlideUrl())
-                    .into(novelDetailsImage)
+                .load(novel.imageUrl?.getGlideUrl())
+                .into(novelDetailsImage)
             novelDetailsImage.setOnClickListener { startImagePreviewActivity(novel.imageUrl, novel.imageFilePath, novelDetailsImage) }
         }
     }
 
     private fun setNovelAuthor() {
-        val author = novel.metaData["Author(s)"]
+        val author = novel.metadata["Author(s)"]
         if (author != null) {
             novelDetailsAuthor.movementMethod = TextViewLinkHandler(this)
             @Suppress("DEPRECATION")
             novelDetailsAuthor.applyFont(assets).text = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N)
                 Html.fromHtml(author, Html.FROM_HTML_MODE_LEGACY) else Html.fromHtml(author)
+        }
+    }
+
+    @Suppress("DEPRECATION")
+    private fun setLicensingInfo() {
+        if (novel.metadata["English Publisher"] ?: "" != "" || novel.metadata["Licensed (in English)"] == "Yes") {
+            val publisher = if (novel.metadata["English Publisher"] == null || novel.metadata["English Publisher"] == "") "an unknown publisher" else novel.metadata["English Publisher"]
+            val warningLabel = getString(R.string.licensed_warning, publisher)
+            novelDetailsLicensedAlert.movementMethod = TextViewLinkHandler(this)
+            novelDetailsLicensedAlert.text = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N)
+                Html.fromHtml(warningLabel, Html.FROM_HTML_MODE_LEGACY) else Html.fromHtml(warningLabel)
+            novelDetailsLicensedLayout.visibility = View.VISIBLE
+        } else {
+            novelDetailsLicensedLayout.visibility = View.GONE
         }
     }
 
@@ -152,7 +201,7 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
                 novelDetailsRatingBar.rating = rating
                 ratingText = "(" + String.format("%.1f", rating) + ")"
             } catch (e: Exception) {
-                Log.w("Library Activity", "Rating: " + novel.rating, e)
+                Logs.warning("Library Activity", "Rating: " + novel.rating, e)
             }
             novelDetailsRatingText.text = ratingText
         }
@@ -161,7 +210,7 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
     private fun setNovelAddToLibraryButton() {
         if (novel.id == -1L) {
             resetAddToLibraryButton()
-            novelDetailsDownloadButton.setOnClickListener {
+            novelDetailAddToLibraryButton.setOnClickListener {
                 disableAddToLibraryButton()
                 addNovelToDB()
             }
@@ -171,22 +220,27 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
     private fun addNovelToDB() {
         if (novel.id == -1L) {
             novel.id = dbHelper.insertNovel(novel)
+            NovelSync.getInstance(novel)?.applyAsync(lifecycleScope) { if (dataCenter.getSyncAddNovels(it.host)) it.addNovel(novel, null) }
+            firebaseAnalytics.logEvent(FAC.Event.ADD_NOVEL) {
+                param(FAC.Param.NOVEL_NAME, novel.name)
+                param(FAC.Param.NOVEL_URL, novel.url)
+            }
         }
     }
 
     private fun resetAddToLibraryButton() {
-        novelDetailsDownloadButton.setText(getString(R.string.add_to_library))
-        novelDetailsDownloadButton.setIconResource(R.drawable.ic_library_add_white_vector)
-        novelDetailsDownloadButton.setBackgroundColor(ContextCompat.getColor(this@NovelDetailsActivity, android.R.color.transparent))
-        novelDetailsDownloadButton.isClickable = true
+        novelDetailAddToLibraryButton.setText(getString(R.string.add_to_library))
+        novelDetailAddToLibraryButton.setIconResource(R.drawable.ic_library_add_white_vector)
+        novelDetailAddToLibraryButton.setBackgroundColor(ContextCompat.getColor(this@NovelDetailsActivity, android.R.color.transparent))
+        novelDetailAddToLibraryButton.isClickable = true
     }
 
     private fun disableAddToLibraryButton() {
         invalidateOptionsMenu()
-        novelDetailsDownloadButton.setText(getString(R.string.in_library))
-        novelDetailsDownloadButton.setIconResource(R.drawable.ic_local_library_white_vector)
-        novelDetailsDownloadButton.setBackgroundColor(ContextCompat.getColor(this@NovelDetailsActivity, R.color.Green))
-        novelDetailsDownloadButton.isClickable = false
+        novelDetailAddToLibraryButton.setText(getString(R.string.in_library))
+        novelDetailAddToLibraryButton.setIconResource(R.drawable.ic_local_library_white_vector)
+        novelDetailAddToLibraryButton.setBackgroundColor(ContextCompat.getColor(this@NovelDetailsActivity, R.color.Green))
+        novelDetailAddToLibraryButton.isClickable = false
     }
 
 
@@ -224,9 +278,9 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
                 }
             }
 
-            val novelDescription = "${novel.longDescription?.subSequence(0, Math.min(300, novel.longDescription!!.length))}… Expand"
+            val novelDescription = "${novel.longDescription?.subSequence(0, min(300, novel.longDescription?.length ?: 0))}… Expand"
             val ss2 = SpannableString(novelDescription)
-            ss2.setSpan(expandClickable, Math.min(300, novel.longDescription!!.length) + 2, novelDescription.length, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
+            ss2.setSpan(expandClickable, min(300, novel.longDescription?.length ?: 0) + 2, novelDescription.length, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
             novelDetailsDescription.applyFont(assets).text = ss2
             novelDetailsDescription.movementMethod = LinkMovementMethod.getInstance()
         }
@@ -242,26 +296,26 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
         return super.onPrepareOptionsMenu(menu)
     }
 
-    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
-        when {
-            item?.itemId == android.R.id.home -> finish()
-            item?.itemId == R.id.action_delete_novel -> confirmNovelDelete()
-            item?.itemId == R.id.action_share -> shareUrl(novel.url)
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            android.R.id.home -> finish()
+            R.id.action_delete_novel -> confirmNovelDelete()
+            R.id.action_share -> shareUrl(novel.url)
         }
         return super.onOptionsItemSelected(item)
     }
 
     private fun confirmNovelDelete() {
         MaterialDialog.Builder(this)
-                .title(getString(R.string.confirm_remove))
-                .content(getString(R.string.confirm_remove_description_novel))
-                .positiveText(getString(R.string.remove))
-                .negativeText(getString(R.string.cancel))
-                .icon(ContextCompat.getDrawable(this, R.drawable.ic_delete_white_vector)!!)
-                .typeface("source_sans_pro_regular.ttf", "source_sans_pro_regular.ttf")
-                .theme(Theme.DARK)
-                .onPositive { _, _ -> deleteNovel() }
-                .show()
+            .title(getString(R.string.confirm_remove))
+            .content(getString(R.string.confirm_remove_description_novel))
+            .positiveText(getString(R.string.remove))
+            .negativeText(getString(R.string.cancel))
+            .icon(ContextCompat.getDrawable(this, R.drawable.ic_delete_white_vector)!!)
+            .typeface("source_sans_pro_regular.ttf", "source_sans_pro_regular.ttf")
+            .theme(Theme.DARK)
+            .onPositive { _, _ -> deleteNovel() }
+            .show()
     }
 
     private fun deleteNovel() {
@@ -269,6 +323,10 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
         novel.id = -1L
         setNovelAddToLibraryButton()
         invalidateOptionsMenu()
+        firebaseAnalytics.logEvent(FAC.Event.REMOVE_NOVEL) {
+            param(FAC.Param.NOVEL_NAME, novel.name)
+            param(FAC.Param.NOVEL_URL, novel.url)
+        }
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
@@ -290,15 +348,20 @@ class NovelDetailsActivity : BaseActivity(), TextViewLinkHandler.OnClickListener
     }
 
     private fun addNovelToHistory() {
-        val history = dataCenter.loadNovelHistory()
-        history.removeAll { novel.name == it.name }
-        history.add(novel)
-        dataCenter.saveNovelHistory(history)
+        try {
+            var history = dbHelper.getLargePreference(Constants.LargePreferenceKeys.RVN_HISTORY)
+                ?: "[]"
+            var historyList: ArrayList<Novel> = Gson().fromJson(history, object : TypeToken<ArrayList<Novel>>() {}.type)
+            Logs.info(TAG, "Novel Search History Size: ${historyList.size}")
+            historyList.removeAll { novel.name == it.name }
+            if (historyList.size > 99)
+                historyList = ArrayList(historyList.take(99))
+            historyList.add(novel)
+            history = Gson().toJson(historyList)
+            dbHelper.createOrUpdateLargePreference(Constants.LargePreferenceKeys.RVN_HISTORY, history)
+        } catch (e: Exception) {
+            Logs.error(TAG, "Error adding novel to history. Resetting the history", e)
+            dbHelper.deleteLargePreference(Constants.LargePreferenceKeys.RVN_HISTORY)
+        }
     }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        async.cancelAll()
-    }
-
 }
