@@ -1,30 +1,39 @@
 package io.github.gmathi.novellibrary.activity
 
-import android.graphics.Rect
+import android.annotation.SuppressLint
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.support.v4.view.MotionEventCompat
-import android.support.v7.widget.DividerItemDecoration
-import android.support.v7.widget.PopupMenu
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.helper.ItemTouchHelper
 import android.view.Menu
 import android.view.MenuItem
 import android.view.MotionEvent
 import android.view.View
+import androidx.appcompat.widget.PopupMenu
+import androidx.core.content.ContextCompat
+import androidx.core.view.MotionEventCompat
+import androidx.lifecycle.lifecycleScope
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.ItemTouchHelper
 import com.afollestad.materialdialogs.MaterialDialog
+import com.google.firebase.analytics.ktx.logEvent
 import io.github.gmathi.novellibrary.R
 import io.github.gmathi.novellibrary.adapter.GenericAdapter
+import io.github.gmathi.novellibrary.dataCenter
 import io.github.gmathi.novellibrary.database.*
 import io.github.gmathi.novellibrary.dbHelper
-import io.github.gmathi.novellibrary.model.NovelSection
-import io.github.gmathi.novellibrary.util.SimpleItemTouchHelperCallback
-import io.github.gmathi.novellibrary.util.SimpleItemTouchListener
+import io.github.gmathi.novellibrary.extensions.FAC
+import io.github.gmathi.novellibrary.extensions.showEmpty
+import io.github.gmathi.novellibrary.extensions.showLoading
+import io.github.gmathi.novellibrary.model.database.NovelSection
+import io.github.gmathi.novellibrary.network.sync.NovelSync
+import io.github.gmathi.novellibrary.util.view.CustomDividerItemDecoration
+import io.github.gmathi.novellibrary.util.view.SimpleItemTouchHelperCallback
+import io.github.gmathi.novellibrary.util.view.SimpleItemTouchListener
 import io.github.gmathi.novellibrary.util.setDefaults
 import kotlinx.android.synthetic.main.activity_novel_sections.*
 import kotlinx.android.synthetic.main.content_recycler_view.*
 import kotlinx.android.synthetic.main.listitem_novel_section.view.*
-import org.jsoup.helper.StringUtil
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
 class NovelSectionsActivity : BaseActivity(), GenericAdapter.Listener<NovelSection>, SimpleItemTouchListener {
 
@@ -47,17 +56,7 @@ class NovelSectionsActivity : BaseActivity(), GenericAdapter.Listener<NovelSecti
         touchHelper = ItemTouchHelper(callback)
         touchHelper.attachToRecyclerView(recyclerView)
         recyclerView.setDefaults(adapter)
-        recyclerView.addItemDecoration(object : DividerItemDecoration(this, DividerItemDecoration.VERTICAL) {
-
-            override fun getItemOffsets(outRect: Rect?, view: View?, parent: RecyclerView?, state: RecyclerView.State?) {
-                val position = parent?.getChildAdapterPosition(view)
-                if (position == parent?.adapter?.itemCount?.minus(1)) {
-                    outRect?.setEmpty()
-                } else {
-                    super.getItemOffsets(outRect, view, parent, state)
-                }
-            }
-        })
+        recyclerView.addItemDecoration(CustomDividerItemDecoration(this, DividerItemDecoration.VERTICAL))
         swipeRefreshLayout.isEnabled = false
     }
 
@@ -70,11 +69,14 @@ class NovelSectionsActivity : BaseActivity(), GenericAdapter.Listener<NovelSecti
             progressLayout.showContent()
         }
         if (novelSections.isEmpty())
-            progressLayout.showError(ContextCompat.getDrawable(this@NovelSectionsActivity, R.drawable.ic_info_white_vector),
-                    getString(R.string.novel_section_empty_message),
-                    getString(R.string.add_novel_section)) {
-                addNovelSection()
-            }
+            progressLayout.showEmpty(
+                resId = R.drawable.ic_info_white_vector,
+                isLottieAnimation = false,
+                emptyText = getString(R.string.novel_section_empty_message),
+                buttonText = getString(R.string.add_novel_section),
+                onClickListener = View.OnClickListener {
+                    addNovelSection()
+                })
     }
 
     private fun updateOrderIds() {
@@ -85,13 +87,15 @@ class NovelSectionsActivity : BaseActivity(), GenericAdapter.Listener<NovelSecti
     }
 
 
+    @SuppressLint("ClickableViewAccessibility")
     override fun bind(item: NovelSection, itemView: View, position: Int) {
         itemView.novelSectionTitle.text = item.name
 
         itemView.reorderButton.setOnTouchListener { _, event ->
             @Suppress("DEPRECATION")
             if (MotionEventCompat.getActionMasked(event) ==
-                    MotionEvent.ACTION_DOWN) {
+                MotionEvent.ACTION_DOWN
+            ) {
                 touchHelper.startDrag(recyclerView.getChildViewHolder(itemView))
             }
             false
@@ -101,8 +105,8 @@ class NovelSectionsActivity : BaseActivity(), GenericAdapter.Listener<NovelSecti
             val popup = PopupMenu(this@NovelSectionsActivity, it)
             popup.menuInflater.inflate(R.menu.menu_popup_novel_section, popup.menu)
 
-            popup.setOnMenuItemClickListener {
-                when (it.itemId) {
+            popup.setOnMenuItemClickListener { menuItem ->
+                when (menuItem.itemId) {
                     R.id.action_novel_section_rename -> {
                         onItemRename(position)
                         true
@@ -119,11 +123,13 @@ class NovelSectionsActivity : BaseActivity(), GenericAdapter.Listener<NovelSecti
             popup.show()
         }
 
-        itemView.setBackgroundColor(if (position % 2 == 0) ContextCompat.getColor(this, R.color.black_transparent)
-        else ContextCompat.getColor(this, android.R.color.transparent))
+        itemView.setBackgroundColor(
+            if (position % 2 == 0) ContextCompat.getColor(this, R.color.black_transparent)
+            else ContextCompat.getColor(this, android.R.color.transparent)
+        )
     }
 
-    override fun onItemClick(item: NovelSection) {
+    override fun onItemClick(item: NovelSection, position: Int) {
         //Do Nothing
     }
 
@@ -133,9 +139,9 @@ class NovelSectionsActivity : BaseActivity(), GenericAdapter.Listener<NovelSecti
         return super.onCreateOptionsMenu(menu)
     }
 
-    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
-        if (item?.itemId == android.R.id.home) finish()
-        else if (item?.itemId == R.id.action_add_novel_section) addNovelSection()
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        if (item.itemId == android.R.id.home) finish()
+        else if (item.itemId == R.id.action_add_novel_section) addNovelSection()
         return super.onOptionsItemSelected(item)
     }
 
@@ -149,62 +155,78 @@ class NovelSectionsActivity : BaseActivity(), GenericAdapter.Listener<NovelSecti
 
     private fun onItemRemove(position: Int) {
         MaterialDialog.Builder(this)
-                .title(getString(R.string.confirm_remove))
-                .content(getString(R.string.confirm_remove_description_novel_section))
-                .positiveText(R.string.remove)
-                .negativeText(R.string.cancel)
-                .onPositive { dialog, _ ->
-                    run {
-                        val novelSection = adapter.items[position]
-                        if (novelSection.id != -1L) {
-                            dbHelper.getAllNovels(novelSection.id).forEach {
-                                dbHelper.updateNovelSectionId(it.id, -1L)
-                            }
-                            dbHelper.deleteNovelSection(novelSection.id)
+            .title(getString(R.string.confirm_remove))
+            .content(getString(R.string.confirm_remove_description_novel_section))
+            .positiveText(R.string.remove)
+            .negativeText(R.string.cancel)
+            .onPositive { dialog, _ ->
+                run {
+                    val novelSection = adapter.items[position]
+                    if (novelSection.id != -1L) {
+                        dbHelper.getAllNovels(novelSection.id).forEach {
+                            dbHelper.updateNovelSectionId(it.id, -1L)
+                            NovelSync.getInstance(it)?.applyAsync(lifecycleScope) { sync -> if (dataCenter.getSyncAddNovels(sync.host)) sync.updateNovel(it, null) }
                         }
-                        setData()
-                        dialog.dismiss()
+                        dbHelper.deleteNovelSection(novelSection.id)
                     }
-                }
-                .onNegative { dialog, _ ->
-                    run {
-                        adapter.notifyDataSetChanged()
-                        dialog.dismiss()
+                    setData()
+                    firebaseAnalytics.logEvent(FAC.Event.REMOVE_NOVEL_SECTION) {
+                        param(FAC.Param.NOVEL_SECTION_NAME, novelSection.name ?: "N/A")
                     }
+                    dialog.dismiss()
                 }
-                .show()
+            }
+            .onNegative { dialog, _ ->
+                run {
+                    adapter.notifyDataSetChanged()
+                    dialog.dismiss()
+                }
+            }
+            .show()
     }
 
     private fun onItemRename(position: Int) {
         val novelSection = adapter.items[position]
         MaterialDialog.Builder(this)
-                .title(getString(R.string.rename))
-                .input(getString(R.string.novel_section_name), novelSection.name) { dialog, input ->
-                    val newName = input.trim().toString()
-                    if (!StringUtil.isBlank(newName) && dbHelper.getNovelSection(newName) == null) {
-                        dbHelper.updateNovelSectionName(novelSection.id, newName)
-                        setData()
-                    } else {
-                        MaterialDialog.Builder(this@NovelSectionsActivity).content(getString(R.string.novel_section_name_error)).show()
+            .title(getString(R.string.rename))
+            .input(getString(R.string.novel_section_name), novelSection.name) { dialog, input ->
+                val newName = input.trim().toString()
+                if (newName.isNotBlank() && dbHelper.getNovelSection(newName) == null) {
+                    dbHelper.updateNovelSectionName(novelSection.id, newName)
+                    setData()
+                    val oldName = novelSection.name
+                    lifecycleScope.launch {
+                        NovelSync.getAllInstances().forEach {
+                            withContext(Dispatchers.IO) { if (dataCenter.getSyncAddNovels(it.host)) it.renameSection(novelSection, oldName, newName) }
+                        }
                     }
-                    dialog.dismiss()
-                }.show()
+                    firebaseAnalytics.logEvent(FAC.Event.RENAME_NOVEL_SECTION) {
+                        param(FAC.Param.NOVEL_SECTION_NAME, newName)
+                    }
+                } else {
+                    MaterialDialog.Builder(this@NovelSectionsActivity).content(getString(R.string.novel_section_name_error)).show()
+                }
+                dialog.dismiss()
+            }.show()
     }
 
     private fun addNovelSection() {
         MaterialDialog.Builder(this)
-                .title(getString(R.string.add_novel_section_title))
-                .content(getString(R.string.add_novel_section_description))
-                .input(getString(R.string.novel_section_name), "") { dialog, input ->
-                    val name = input.trim().toString()
-                    if (!StringUtil.isBlank(name) && dbHelper.getNovelSection(name) == null) {
-                        dbHelper.createNovelSection(name)
-                        setData()
-                    } else {
-                        MaterialDialog.Builder(this@NovelSectionsActivity).content(getString(R.string.novel_section_name_error)).show()
+            .title(getString(R.string.add_novel_section_title))
+            .content(getString(R.string.add_novel_section_description))
+            .input(getString(R.string.novel_section_name), "") { dialog, input ->
+                val name = input.trim().toString()
+                if (name.isNotBlank() && dbHelper.getNovelSection(name) == null) {
+                    dbHelper.createNovelSection(name)
+                    setData()
+                    firebaseAnalytics.logEvent(FAC.Event.ADD_NOVEL_SECTION) {
+                        param(FAC.Param.NOVEL_SECTION_NAME, name)
                     }
-                    dialog.dismiss()
-                }.show()
+                } else {
+                    MaterialDialog.Builder(this@NovelSectionsActivity).content(getString(R.string.novel_section_name_error)).show()
+                }
+                dialog.dismiss()
+            }.show()
     }
 
     override fun onPause() {
