@@ -2,12 +2,12 @@ package io.github.gmathi.novellibrary.activity
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.webkit.WebViewClient
 import io.github.gmathi.novellibrary.R
+import io.github.gmathi.novellibrary.network.HostNames
 import kotlinx.android.synthetic.main.activity_web_view.*
 
-class WebViewActivity : AppCompatActivity() {
+class WebViewActivity : BaseActivity() {
 
     @SuppressLint("SetJavaScriptEnabled")
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -17,9 +17,11 @@ class WebViewActivity : AppCompatActivity() {
         previewWebView.apply {
             settings.apply {
                 javaScriptEnabled = true
+                userAgentString = HostNames.USER_AGENT
             }
             webViewClient = WebViewClient()
-            loadUrl(intent.getStringExtra("url"))
+            val url = intent.getStringExtra("url") ?: return
+            loadUrl(url)
         }
     }
 }
