@@ -1,16 +1,11 @@
 package io.github.gmathi.novellibrary.cleaner
 
-import android.net.Uri
 import org.jsoup.nodes.Document
-import org.jsoup.nodes.Element
-import java.io.File
 
-class TumblrCleaner : HtmlHelper() {
+class TumblrCleaner : HtmlCleaner() {
 
     override fun additionalProcessing(doc: Document) {
         removeCSS(doc)
-        doc.getElementsByTag("link")?.remove()
-        doc.getElementsByTag("style")?.remove()
         var contentElement = doc.body().getElementsByTag("div").firstOrNull { it.hasClass("textpostbody") }
 
         if (contentElement == null) {
@@ -26,16 +21,5 @@ class TumblrCleaner : HtmlHelper() {
         cleanClassAndIds(contentElement)
         contentElement?.getElementsByClass("wpcnt")?.remove()
         contentElement?.getElementById("jp-post-flair")?.remove()
-    }
-
-    override fun downloadImage(element: Element, dir: File): File? {
-        val uri = Uri.parse(element.attr("src"))
-        return if (uri.toString().contains("uploads/avatars")) null
-        else super.downloadImage(element, dir)
-    }
-
-
-    override fun toggleTheme(isDark: Boolean, doc: Document): Document {
-        return super.toggleThemeDefault(isDark, doc)
     }
 }
