@@ -1,11 +1,15 @@
 package io.github.gmathi.novellibrary.database
 
 import android.content.ContentValues
-import android.util.Log
+import android.database.Cursor
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import io.github.gmathi.novellibrary.model.Novel
-import io.github.gmathi.novellibrary.model.NovelGenre
+import io.github.gmathi.novellibrary.model.database.Novel
+import io.github.gmathi.novellibrary.model.database.NovelGenre
+import io.github.gmathi.novellibrary.network.NovelApi
+import io.github.gmathi.novellibrary.network.getNovelDetails
+import io.github.gmathi.novellibrary.util.Logs
+import kotlinx.coroutines.coroutineScope
 import java.util.*
 
 private const val LOG = "NovelHelper"
@@ -27,7 +31,7 @@ fun DBHelper.createNovel(novel: Novel): Long {
     val values = ContentValues()
     values.put(DBKeys.KEY_NAME, novel.name)
     values.put(DBKeys.KEY_URL, novel.url)
-    values.put(DBKeys.KEY_METADATA, Gson().toJson(novel.metaData))
+    values.put(DBKeys.KEY_METADATA, Gson().toJson(novel.metadata))
     values.put(DBKeys.KEY_IMAGE_URL, novel.imageUrl)
     values.put(DBKeys.KEY_RATING, novel.rating)
     values.put(DBKeys.KEY_SHORT_DESCRIPTION, novel.shortDescription)
@@ -35,46 +39,34 @@ fun DBHelper.createNovel(novel: Novel): Long {
     values.put(DBKeys.KEY_IMAGE_FILE_PATH, novel.imageFilePath)
     values.put(DBKeys.KEY_NEW_RELEASES_COUNT, novel.newReleasesCount)
     values.put(DBKeys.KEY_CHAPTERS_COUNT, novel.chaptersCount)
-    values.put(DBKeys.KEY_CURRENT_WEB_PAGE_ID, novel.currentWebPageId)
+    values.put(DBKeys.KEY_CURRENT_WEB_PAGE_URL, novel.currentChapterUrl)
     values.put(DBKeys.KEY_NOVEL_SECTION_ID, novel.novelSectionId)
 
     return db.insert(DBKeys.TABLE_NOVEL, null, values)
 }
 
 fun DBHelper.getNovel(novelName: String): Novel? {
-    val selectQuery = "SELECT * FROM " + DBKeys.TABLE_NOVEL + " WHERE " + DBKeys.KEY_NAME + " = \"" + novelName + "\""
-    return getNovelFromQuery(selectQuery)
+    val selectQuery = "SELECT * FROM ${DBKeys.TABLE_NOVEL} WHERE ${DBKeys.KEY_NAME} = ?"
+    return getNovelFromQuery(selectQuery, arrayOf(novelName))
 }
 
 fun DBHelper.getNovelByUrl(novelUrl: String): Novel? {
-    val selectQuery = "SELECT * FROM " + DBKeys.TABLE_NOVEL + " WHERE " + DBKeys.KEY_URL + " = \"" + novelUrl + "\""
-    return getNovelFromQuery(selectQuery)
+    val selectQuery = "SELECT * FROM ${DBKeys.TABLE_NOVEL} WHERE ${DBKeys.KEY_URL} = ?"
+    return getNovelFromQuery(selectQuery, arrayOf(novelUrl))
 }
 
 fun DBHelper.getNovel(novelId: Long): Novel? {
-    val selectQuery = "SELECT * FROM " + DBKeys.TABLE_NOVEL + " WHERE " + DBKeys.KEY_ID + " = " + novelId
-    return getNovelFromQuery(selectQuery)
+    val selectQuery = "SELECT * FROM ${DBKeys.TABLE_NOVEL} WHERE ${DBKeys.KEY_ID} = ?"
+    return getNovelFromQuery(selectQuery, arrayOf(novelId.toString()))
 }
 
-fun DBHelper.getNovelFromQuery(selectQuery: String): Novel? {
-    Log.d(LOG, selectQuery)
-    val db = this.readableDatabase
-    val cursor = db.rawQuery(selectQuery, null)
+fun DBHelper.getNovelFromQuery(selectQuery: String, selectionArgs: Array<String>? = null): Novel? {
+    Logs.debug(LOG, selectQuery)
     var novel: Novel? = null
+    val cursor = this.readableDatabase.rawQuery(selectQuery, selectionArgs)
     if (cursor != null) {
         if (cursor.moveToFirst()) {
-            novel = Novel(cursor.getString(cursor.getColumnIndex(DBKeys.KEY_NAME)), cursor.getString(cursor.getColumnIndex(DBKeys.KEY_URL)))
-            novel.id = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_ID))
-            novel.metaData = Gson().fromJson(cursor.getString(cursor.getColumnIndex(DBKeys.KEY_METADATA)), object : TypeToken<HashMap<String, String>>() {}.type)
-            novel.imageUrl = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_IMAGE_URL))
-            novel.rating = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_RATING))
-            novel.shortDescription = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_SHORT_DESCRIPTION))
-            novel.longDescription = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_LONG_DESCRIPTION))
-            novel.imageFilePath = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_IMAGE_FILE_PATH))
-            novel.newReleasesCount = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_NEW_RELEASES_COUNT))
-            novel.chaptersCount = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_CHAPTERS_COUNT))
-            novel.currentWebPageId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_CURRENT_WEB_PAGE_ID))
-            novel.novelSectionId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_NOVEL_SECTION_ID))
+            novel = getNovelFromCursor(cursor)
         }
         cursor.close()
     }
@@ -83,13 +75,62 @@ fun DBHelper.getNovelFromQuery(selectQuery: String): Novel? {
     return novel
 }
 
-fun DBHelper.getNovelId(novelName: String): Long {
-    val db = this.readableDatabase
-    val selectQuery = "SELECT id FROM " + DBKeys.TABLE_NOVEL + " WHERE " + DBKeys.KEY_NAME + " = \"" + novelName + "\""
+fun getNovelFromCursor(cursor: Cursor): Novel {
+    val novel = Novel(cursor.getString(cursor.getColumnIndex(DBKeys.KEY_NAME)), cursor.getString(cursor.getColumnIndex(DBKeys.KEY_URL)))
+    novel.id = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_ID))
+    novel.metadata = Gson().fromJson(cursor.getString(cursor.getColumnIndex(DBKeys.KEY_METADATA)), object : TypeToken<HashMap<String, String>>() {}.type)
+    novel.imageUrl = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_IMAGE_URL))
+    novel.rating = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_RATING))
+    novel.shortDescription = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_SHORT_DESCRIPTION))
+    novel.longDescription = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_LONG_DESCRIPTION))
+    novel.imageFilePath = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_IMAGE_FILE_PATH))
+    novel.newReleasesCount = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_NEW_RELEASES_COUNT))
+    novel.chaptersCount = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_CHAPTERS_COUNT))
+    novel.currentChapterUrl = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_CURRENT_WEB_PAGE_URL))
+    novel.novelSectionId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_NOVEL_SECTION_ID))
+    return novel
+}
 
-    Log.d(LOG, selectQuery)
+
+fun DBHelper.getAllNovels(): List<Novel> {
+    val selectQuery = "SELECT * FROM novel ORDER BY ${DBKeys.KEY_ORDER_ID} ASC"
+    Logs.debug(LOG, selectQuery)
+    val db = this.readableDatabase
     val cursor = db.rawQuery(selectQuery, null)
-    var id: Long = -1
+    val list = ArrayList<Novel>()
+    if (cursor != null) {
+        if (cursor.moveToFirst()) {
+            do {
+                val novel = getNovelFromCursor(cursor)
+                list.add(novel)
+            } while (cursor.moveToNext())
+        }
+        cursor.close()
+    }
+    return list
+}
+
+fun DBHelper.getAllNovels(novelSectionId: Long): List<Novel> {
+    val selectQuery = "SELECT * FROM novel WHERE ${DBKeys.KEY_NOVEL_SECTION_ID} = $novelSectionId ORDER BY ${DBKeys.KEY_ORDER_ID} ASC"
+    val db = this.readableDatabase
+    val cursor = db.rawQuery(selectQuery, null)
+    val list = ArrayList<Novel>()
+    if (cursor != null) {
+        if (cursor.moveToFirst()) {
+            do {
+                val novel = getNovelFromCursor(cursor)
+                list.add(novel)
+            } while (cursor.moveToNext())
+        }
+        cursor.close()
+    }
+    return list
+}
+
+fun DBHelper.getNovelId(novelName: String): Long {
+    var id = -1L
+    val selectQuery = "SELECT id FROM ${DBKeys.TABLE_NOVEL} WHERE ${DBKeys.KEY_NAME} = ?"
+    val cursor = this.readableDatabase.rawQuery(selectQuery, arrayOf(novelName))
     if (cursor != null) {
         if (cursor.moveToFirst()) {
             id = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_ID))
@@ -99,94 +140,13 @@ fun DBHelper.getNovelId(novelName: String): Long {
     return id
 }
 
-fun DBHelper.getAllNovels(): List<Novel> {
-
-    val selectQuery = "select * from novel ORDER BY " + DBKeys.KEY_ORDER_ID + " ASC"
-
-    Log.d(LOG, selectQuery)
-
-    val db = this.readableDatabase
-    val cursor = db.rawQuery(selectQuery, null)
-
-    val list = ArrayList<Novel>()
-    if (cursor != null) {
-        if (cursor.moveToFirst()) {
-            do {
-                val novel = Novel(cursor.getString(cursor.getColumnIndex(DBKeys.KEY_NAME)), cursor.getString(cursor.getColumnIndex(DBKeys.KEY_URL)))
-                novel.id = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_ID))
-                novel.metaData = Gson().fromJson(cursor.getString(cursor.getColumnIndex(DBKeys.KEY_METADATA)), object : TypeToken<HashMap<String, String>>() {}.type)
-                novel.imageUrl = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_IMAGE_URL))
-                novel.rating = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_RATING))
-                novel.shortDescription = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_SHORT_DESCRIPTION))
-                novel.longDescription = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_LONG_DESCRIPTION))
-                novel.imageFilePath = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_IMAGE_FILE_PATH))
-                novel.currentWebPageId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_CURRENT_WEB_PAGE_ID))
-                novel.newReleasesCount = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_NEW_RELEASES_COUNT))
-                novel.chaptersCount = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_CHAPTERS_COUNT))
-                novel.orderId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_ORDER_ID))
-                novel.novelSectionId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_NOVEL_SECTION_ID))
-                list.add(novel)
-            } while (cursor.moveToNext())
-        }
-        cursor.close()
-    }
-
-    list.forEach {
-
-    }
-
-
-    return list
-}
-
-fun DBHelper.getAllNovels(novelSectionId: Long): List<Novel> {
-
-    val selectQuery = "SELECT * FROM novel WHERE ${DBKeys.KEY_NOVEL_SECTION_ID} = $novelSectionId ORDER BY ${DBKeys.KEY_ORDER_ID} ASC"
-
-    Log.d(LOG, selectQuery)
-
-    val db = this.readableDatabase
-    val cursor = db.rawQuery(selectQuery, null)
-
-    val list = ArrayList<Novel>()
-    if (cursor != null) {
-        if (cursor.moveToFirst()) {
-            do {
-                val novel = Novel(cursor.getString(cursor.getColumnIndex(DBKeys.KEY_NAME)), cursor.getString(cursor.getColumnIndex(DBKeys.KEY_URL)))
-                novel.id = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_ID))
-                novel.metaData = Gson().fromJson(cursor.getString(cursor.getColumnIndex(DBKeys.KEY_METADATA)), object : TypeToken<HashMap<String, String>>() {}.type)
-                novel.imageUrl = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_IMAGE_URL))
-                novel.rating = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_RATING))
-                novel.shortDescription = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_SHORT_DESCRIPTION))
-                novel.longDescription = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_LONG_DESCRIPTION))
-                novel.imageFilePath = cursor.getString(cursor.getColumnIndex(DBKeys.KEY_IMAGE_FILE_PATH))
-                novel.currentWebPageId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_CURRENT_WEB_PAGE_ID))
-                novel.newReleasesCount = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_NEW_RELEASES_COUNT))
-                novel.chaptersCount = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_CHAPTERS_COUNT))
-                novel.orderId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_ORDER_ID))
-                novel.novelSectionId = cursor.getLong(cursor.getColumnIndex(DBKeys.KEY_NOVEL_SECTION_ID))
-                list.add(novel)
-            } while (cursor.moveToNext())
-        }
-        cursor.close()
-    }
-
-    list.forEach {
-
-    }
-
-
-    return list
-}
 
 fun DBHelper.updateNovel(novel: Novel): Long {
     val values = ContentValues()
-
-    //values.put(DBKeys.KEY_ID, novel.id)
     values.put(DBKeys.KEY_NAME, novel.name)
     values.put(DBKeys.KEY_URL, novel.url)
-    if (novel.metaData.isNotEmpty())
-        values.put(DBKeys.KEY_METADATA, Gson().toJson(novel.metaData))
+    if (novel.metadata.isNotEmpty())
+        values.put(DBKeys.KEY_METADATA, Gson().toJson(novel.metadata))
     values.put(DBKeys.KEY_IMAGE_URL, novel.imageUrl)
     values.put(DBKeys.KEY_RATING, novel.rating)
     values.put(DBKeys.KEY_SHORT_DESCRIPTION, novel.shortDescription)
@@ -198,8 +158,8 @@ fun DBHelper.updateNovel(novel: Novel): Long {
     values.put(DBKeys.KEY_LONG_DESCRIPTION, novel.longDescription)
     if (novel.imageFilePath != null)
         values.put(DBKeys.KEY_IMAGE_FILE_PATH, novel.imageFilePath)
-    if (novel.currentWebPageId != -1L)
-        values.put(DBKeys.KEY_CURRENT_WEB_PAGE_ID, novel.currentWebPageId)
+    if (novel.currentChapterUrl != null)
+        values.put(DBKeys.KEY_CURRENT_WEB_PAGE_URL, novel.currentChapterUrl)
     if (novel.novelSectionId != -1L)
         values.put(DBKeys.KEY_NOVEL_SECTION_ID, novel.novelSectionId)
     if (novel.genres != null) {
@@ -225,9 +185,9 @@ fun DBHelper.updateNovelSectionId(novelId: Long, novelSectionId: Long) {
     this.writableDatabase.update(DBKeys.TABLE_NOVEL, values, DBKeys.KEY_ID + " = ?", arrayOf(novelId.toString())).toLong()
 }
 
-fun DBHelper.updateBookmarkCurrentWebPageId(novelId: Long, currentWebPageId: Long?) {
+fun DBHelper.updateBookmarkCurrentWebPageUrl(novelId: Long, currentChapterUrl: String?) {
     val values = ContentValues()
-    values.put(DBKeys.KEY_CURRENT_WEB_PAGE_ID, currentWebPageId)
+    values.put(DBKeys.KEY_CURRENT_WEB_PAGE_URL, currentChapterUrl)
     this.writableDatabase.update(DBKeys.TABLE_NOVEL, values, DBKeys.KEY_ID + " = ?", arrayOf(novelId.toString())).toLong()
 }
 
@@ -245,7 +205,7 @@ fun DBHelper.updateNewReleasesCount(novelId: Long, newReleasesCount: Long) {
 
 fun DBHelper.updateNovelMetaData(novel: Novel) {
     val values = ContentValues()
-    values.put(DBKeys.KEY_METADATA, Gson().toJson(novel.metaData))
+    values.put(DBKeys.KEY_METADATA, Gson().toJson(novel.metadata))
     this.writableDatabase.update(DBKeys.TABLE_NOVEL, values, DBKeys.KEY_ID + " = ?", arrayOf(novel.id.toString())).toLong()
 }
 
@@ -256,8 +216,24 @@ fun DBHelper.updateChaptersAndReleasesCount(novelId: Long, totalChaptersCount: L
     this.writableDatabase.update(DBKeys.TABLE_NOVEL, values, DBKeys.KEY_ID + " = ?", arrayOf(novelId.toString())).toLong()
 }
 
-
 fun DBHelper.deleteNovel(id: Long) {
     this.writableDatabase.delete(DBKeys.TABLE_NOVEL, DBKeys.KEY_ID + " = ?", arrayOf(id.toString()))
 }
 
+fun DBHelper.updateChaptersCount(novelId: Long, chaptersCount: Long) {
+    val values = ContentValues()
+    values.put(DBKeys.KEY_CHAPTERS_COUNT, chaptersCount)
+    this.writableDatabase.update(DBKeys.TABLE_NOVEL, values, DBKeys.KEY_ID + " = ?", arrayOf(novelId.toString())).toLong()
+}
+
+
+suspend fun DBHelper.resetNovel(novel: Novel) = coroutineScope {
+    // Completely delete all novel data and start fresh. Hard reset mode ;p
+    cleanupNovelData(novel)
+
+    // Notice: Cannot run getNovelDetails on MainThread
+    val newNovel = NovelApi.getNovelDetails(novel.url)
+    newNovel?.novelSectionId =  novel.novelSectionId
+    newNovel?.orderId = novel.orderId
+    if (newNovel != null) insertNovel(newNovel)
+}
