@@ -1,8 +1,8 @@
 package io.github.gmathi.novellibrary.database
 
 import android.content.ContentValues
-import android.util.Log
-import io.github.gmathi.novellibrary.model.NovelSection
+import io.github.gmathi.novellibrary.model.database.NovelSection
+import io.github.gmathi.novellibrary.util.Logs
 
 
 private const val LOG = "NovelSectionHelper"
@@ -17,8 +17,8 @@ fun DBHelper.createNovelSection(novelSectionName: String): Long {
 }
 
 fun DBHelper.getNovelSection(novelSectionName: String): NovelSection? {
-    val selectQuery = "SELECT * FROM ${DBKeys.TABLE_NOVEL_SECTION} WHERE ${DBKeys.KEY_NAME} = \"$novelSectionName\""
-    return getNovelSectionFromQuery(selectQuery)
+    val selectQuery = "SELECT * FROM ${DBKeys.TABLE_NOVEL_SECTION} WHERE ${DBKeys.KEY_NAME} = ?"
+    return getNovelSectionFromQuery(selectQuery, arrayOf(novelSectionName))
 }
 
 fun DBHelper.getNovelSection(novelSectionId: Long): NovelSection? {
@@ -26,10 +26,10 @@ fun DBHelper.getNovelSection(novelSectionId: Long): NovelSection? {
     return getNovelSectionFromQuery(selectQuery)
 }
 
-fun DBHelper.getNovelSectionFromQuery(selectQuery: String): NovelSection? {
+fun DBHelper.getNovelSectionFromQuery(selectQuery: String, selectionArgs: Array<String>? = null): NovelSection? {
     val db = this.readableDatabase
-    Log.d(LOG, selectQuery)
-    val cursor = db.rawQuery(selectQuery, null)
+    Logs.debug(LOG, selectQuery)
+    val cursor = db.rawQuery(selectQuery, selectionArgs)
     var novelSection: NovelSection? = null
     if (cursor != null) {
         if (cursor.moveToFirst()) {
@@ -46,7 +46,7 @@ fun DBHelper.getNovelSectionFromQuery(selectQuery: String): NovelSection? {
 fun DBHelper.getAllNovelSections(): List<NovelSection> {
     val list = ArrayList<NovelSection>()
     val selectQuery = "SELECT  * FROM ${DBKeys.TABLE_NOVEL_SECTION} ORDER BY ${DBKeys.KEY_ORDER_ID} ASC"
-    Log.d(LOG, selectQuery)
+    Logs.debug(LOG, selectQuery)
     val db = this.readableDatabase
     val cursor = db.rawQuery(selectQuery, null)
     if (cursor != null) {
@@ -90,8 +90,10 @@ fun DBHelper.updateNovelSectionName(novelSectionId: Long, name: String) {
 
 fun DBHelper.deleteNovelSection(id: Long) {
     val db = this.writableDatabase
-    db.delete(DBKeys.TABLE_NOVEL_SECTION, DBKeys.KEY_ID + " = ?",
-            arrayOf(id.toString()))
+    db.delete(
+        DBKeys.TABLE_NOVEL_SECTION, DBKeys.KEY_ID + " = ?",
+        arrayOf(id.toString())
+    )
 }
 
 
