@@ -2,20 +2,28 @@ package io.github.gmathi.novellibrary.fragment
 
 import android.animation.Animator
 import android.os.Bundle
-import android.support.v4.view.GravityCompat
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.core.view.GravityCompat
+import com.google.firebase.analytics.ktx.logEvent
 import io.github.gmathi.novellibrary.R
-import io.github.gmathi.novellibrary.adapter.*
+import io.github.gmathi.novellibrary.activity.BaseActivity
+import io.github.gmathi.novellibrary.adapter.GenericAdapter
+import io.github.gmathi.novellibrary.adapter.GenericFragmentStatePagerAdapter
+import io.github.gmathi.novellibrary.adapter.NavPageListener
+import io.github.gmathi.novellibrary.adapter.SearchResultsListener
 import io.github.gmathi.novellibrary.dataCenter
-import io.github.gmathi.novellibrary.model.Novel
-import io.github.gmathi.novellibrary.util.SimpleAnimationListener
-import io.github.gmathi.novellibrary.util.SuggestionsBuilder
-import io.github.gmathi.novellibrary.util.addToSearchHistory
+import io.github.gmathi.novellibrary.extensions.FAC
+import io.github.gmathi.novellibrary.model.database.Novel
+import io.github.gmathi.novellibrary.util.view.SimpleAnimationListener
+import io.github.gmathi.novellibrary.util.view.SuggestionsBuilder
+import io.github.gmathi.novellibrary.util.addToNovelSearchHistory
+import io.github.gmathi.novellibrary.util.system.hideSoftKeyboard
 import kotlinx.android.synthetic.main.activity_nav_drawer.*
 import kotlinx.android.synthetic.main.fragment_search.*
 import org.cryse.widget.persistentsearch.PersistentSearchView
+import org.cryse.widget.persistentsearch.SearchItem
 
 
 class SearchFragment : BaseFragment() {
@@ -30,7 +38,7 @@ class SearchFragment : BaseFragment() {
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
-            inflater.inflate(R.layout.fragment_search, container, false)
+        inflater.inflate(R.layout.fragment_search, container, false)
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
@@ -68,40 +76,43 @@ class SearchFragment : BaseFragment() {
             hideSoftKeyboard()
             activity?.drawerLayout?.openDrawer(GravityCompat.START)
         }
-        searchView.setSuggestionBuilder(SuggestionsBuilder())
+
+        searchView.setSuggestionBuilder(SuggestionsBuilder(dataCenter.loadNovelSearchHistory()))
         searchView.setSearchListener(object : PersistentSearchView.SearchListener {
 
-            override fun onSearch(searchTerm: String?) {
-                searchTerm?.addToSearchHistory()
-                if (searchTerm != null) {
-                    searchNovels(searchTerm)
-                } else {
-                    // Throw a empty search
+            override fun onSearch(query: String?) {
+                query?.addToNovelSearchHistory()
+                if (query != null) {
+                    searchNovels(query)
+                    searchView.setSuggestionBuilder(SuggestionsBuilder(dataCenter.loadNovelSearchHistory()))
+                    (activity as BaseActivity).firebaseAnalytics.logEvent(FAC.Event.SEARCH_NOVEL) {
+                        param(FAC.Param.SEARCH_TERM, query)
+                    }
                 }
             }
 
             override fun onSearchEditOpened() {
                 searchViewBgTint.visibility = View.VISIBLE
                 searchViewBgTint
-                        .animate()
-                        .alpha(1.0f)
-                        .setDuration(300)
-                        .setListener(SimpleAnimationListener())
-                        .start()
+                    .animate()
+                    .alpha(1.0f)
+                    .setDuration(300)
+                    .setListener(SimpleAnimationListener())
+                    .start()
             }
 
             override fun onSearchEditClosed() {
                 searchViewBgTint
-                        .animate()
-                        .alpha(0.0f)
-                        .setDuration(300)
-                        .setListener(object : SimpleAnimationListener() {
-                            override fun onAnimationEnd(animation: Animator) {
-                                super.onAnimationEnd(animation)
-                                searchViewBgTint.visibility = View.GONE
-                            }
-                        })
-                        .start()
+                    .animate()
+                    .alpha(0.0f)
+                    .setDuration(300)
+                    .setListener(object : SimpleAnimationListener() {
+                        override fun onAnimationEnd(animation: Animator) {
+                            super.onAnimationEnd(animation)
+                            searchViewBgTint.visibility = View.GONE
+                        }
+                    })
+                    .start()
             }
 
             override fun onSearchExit() {
@@ -114,8 +125,12 @@ class SearchFragment : BaseFragment() {
                 //Toast.makeText(context, "onSearchCleared", Toast.LENGTH_SHORT).show()
             }
 
-            override fun onSearchTermChanged(searchTerm: String?) {
+            override fun onSearchTermChanged(term: String?) {
                 //Toast.makeText(context, "Search Exited", Toast.LENGTH_SHORT).show()
+            }
+
+            override fun onSuggestion(searchItem: SearchItem?): Boolean {
+                return true
             }
 
             override fun onSearchEditBackPressed(): Boolean {
@@ -136,15 +151,20 @@ class SearchFragment : BaseFragment() {
         searchMode = true
         this.searchTerm = searchTerm
 
-        val titles: Array<out String>
+        val titles = ArrayList<String>()
+        titles.add("Novel-Updates")
+        if (!dataCenter.lockRoyalRoad || dataCenter.isDeveloper)
+            titles.add("RoyalRoad")
+        if (!dataCenter.lockNovelFull || dataCenter.isDeveloper)
+            titles.add("NovelFull")
+        if (!dataCenter.lockScribble || dataCenter.isDeveloper)
+            titles.add("ScribbleHub")
+        titles.add("WLN-Updates")
+        titles.add("LNMTL")
+        titles.add("Neovel")
+
         val searchPageAdapter: GenericFragmentStatePagerAdapter
-        if (dataCenter.lockRoyalRoad) {
-            titles = resources.getStringArray(R.array.search_results_tab_titles)
-            searchPageAdapter = GenericFragmentStatePagerAdapter(childFragmentManager, titles, titles.size, SearchResultsListener(searchTerm))
-        } else {
-            titles = resources.getStringArray(R.array.search_results_tab_titles_unlocked)
-            searchPageAdapter = GenericFragmentStatePagerAdapter(childFragmentManager, titles, titles.size, SearchResultsUnlockedListener(searchTerm))
-        }
+        searchPageAdapter = GenericFragmentStatePagerAdapter(childFragmentManager, titles.toTypedArray(), titles.size, SearchResultsListener(searchTerm, titles))
 
         viewPager.offscreenPageLimit = 2
         viewPager.adapter = searchPageAdapter
