@@ -1,34 +1,46 @@
 package io.github.gmathi.novellibrary.fragment
 
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import co.metalab.asyncawait.async
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.lifecycleScope
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import io.github.gmathi.novellibrary.R
-import io.github.gmathi.novellibrary.activity.startNovelDetailsActivity
 import io.github.gmathi.novellibrary.adapter.GenericAdapter
 import io.github.gmathi.novellibrary.dataCenter
-import io.github.gmathi.novellibrary.model.Novel
+import io.github.gmathi.novellibrary.extensions.*
+import io.github.gmathi.novellibrary.model.database.Novel
 import io.github.gmathi.novellibrary.network.*
+import io.github.gmathi.novellibrary.util.Logs
 import io.github.gmathi.novellibrary.util.Utils
 import io.github.gmathi.novellibrary.util.getGlideUrl
 import io.github.gmathi.novellibrary.util.setDefaults
+import io.github.gmathi.novellibrary.util.system.isFragmentActive
+import io.github.gmathi.novellibrary.util.system.startNovelDetailsActivity
 import kotlinx.android.synthetic.main.content_recycler_view.*
 import kotlinx.android.synthetic.main.listitem_novel.view.*
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import java.net.URLEncoder
+import java.util.concurrent.atomic.AtomicBoolean
 
 
-class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
+class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel>, GenericAdapter.LoadMoreListener {
 
+
+    override var currentPageNumber: Int = 1
+    override val preloadCount: Int = 50
+    override val isPageLoading: AtomicBoolean = AtomicBoolean(false)
     private lateinit var searchTerm: String
     private lateinit var resultType: String
 
     companion object {
+        private const val TAG = "SearchTermFragment"
+
         fun newInstance(searchTerms: String, resultType: String): SearchTermFragment {
             val bundle = Bundle()
             bundle.putString("searchTerm", searchTerms)
@@ -54,8 +66,8 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
         super.onActivityCreated(savedInstanceState)
         //(activity as AppCompatActivity).setSupportActionBar(null)
 
-        searchTerm = arguments!!.getString("searchTerm")
-        resultType = arguments!!.getString("resultType")
+        searchTerm = arguments?.getString("searchTerm")!!
+        resultType = arguments?.getString("resultType")!!
 
         setRecyclerView()
 
@@ -72,23 +84,19 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
     }
 
     private fun setRecyclerView() {
-        adapter = GenericAdapter(items = ArrayList(), layoutResId = R.layout.listitem_novel, listener = this)
+        adapter = GenericAdapter(items = ArrayList(), layoutResId = R.layout.listitem_novel, listener = this, loadMoreListener = if (resultType != HostNames.WLN_UPDATES) this else null)
         recyclerView.setDefaults(adapter)
         swipeRefreshLayout.setOnRefreshListener { searchNovels() }
     }
 
     private fun searchNovels() {
 
-        async search@ {
-
-            if (resultType == HostNames.ROYAL_ROAD && dataCenter.lockRoyalRoad) {
-                progressLayout.showEmpty(ContextCompat.getDrawable(context!!, R.drawable.ic_phonelink_lock_white_vector), getString(R.string.content_restricted))
-                return@search
-            }
+        lifecycleScope.launch search@{
 
             if (!Utils.isConnectedToNetwork(activity)) {
-                progressLayout.showError(ContextCompat.getDrawable(context!!, R.drawable.ic_warning_white_vector), getString(R.string.no_internet), getString(R.string.try_again), {
+                progressLayout.noInternetError(View.OnClickListener {
                     progressLayout.showLoading()
+                    currentPageNumber = 1
                     searchNovels()
                 })
                 return@search
@@ -97,51 +105,95 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
             val searchTerms = URLEncoder.encode(searchTerm, "UTF-8")
             var results: ArrayList<Novel>? = null
 
-            when (resultType) {
-                HostNames.NOVEL_UPDATES -> results = await { NovelApi.searchNovelUpdates(searchTerms) }
-                HostNames.ROYAL_ROAD -> results = await { NovelApi.searchRoyalRoad(searchTerms) }
-                HostNames.WLN_UPDATES -> results = await { NovelApi.searchWlnUpdates(searchTerms) }
+            try {
+                when (resultType) {
+                    HostNames.NOVEL_UPDATES -> results = withContext(Dispatchers.IO) { NovelApi.searchNovelUpdates(searchTerms, currentPageNumber) }
+                    HostNames.ROYAL_ROAD -> results = withContext(Dispatchers.IO) { NovelApi.searchRoyalRoad(searchTerms, currentPageNumber) }
+                    HostNames.NOVEL_FULL -> results = withContext(Dispatchers.IO) { NovelApi.searchNovelFull(searchTerms, currentPageNumber) }
+                    HostNames.WLN_UPDATES -> results = withContext(Dispatchers.IO) { NovelApi.searchWlnUpdates(searchTerms) }
+                    HostNames.SCRIBBLE_HUB -> results = withContext(Dispatchers.IO) { NovelApi.searchScribbleHub(searchTerms, currentPageNumber) }
+                    HostNames.LNMTL -> results = withContext(Dispatchers.IO) { NovelApi.searchLNMTL(searchTerms) }
+                    HostNames.NEOVEL -> results =  withContext(Dispatchers.IO) { NovelApi.searchNeovel(searchTerms) }
+                }
+            } catch (e: Exception) {
+                Logs.error(TAG, "Search Failed!", e)
             }
 
-            if (results == null) results = ArrayList()
-            if (isFragmentActive() && progressLayout != null) {
-                loadSearchResults(results)
-                swipeRefreshLayout.isRefreshing = false
+            if (results != null) {
+                if (isVisible && (!isDetached || !isRemoving)) {
+                    loadSearchResults(results)
+                    isPageLoading.lazySet(false)
+                    swipeRefreshLayout.isRefreshing = false
+                }
+            } else {
+                if (isFragmentActive() && progressLayout != null)
+                    progressLayout.showError(errorText = getString(R.string.connection_error), buttonText = getString(R.string.try_again), onClickListener = View.OnClickListener {
+                        progressLayout.showLoading()
+                        currentPageNumber = 1
+                        searchNovels()
+                    })
             }
         }
     }
 
     private fun loadSearchResults(results: ArrayList<Novel>) {
-        adapter.updateData(results)
-        if (adapter.items.isEmpty()) {
-            progressLayout.showError(ContextCompat.getDrawable(context!!, R.drawable.ic_youtube_searched_for_white_vector), "No Novels Found!", "Try Again", {
-                progressLayout.showLoading()
-                searchNovels()
-            })
+        if (results.isNotEmpty() && !adapter.items.containsAll(results)) {
+            if (currentPageNumber == 1) {
+                adapter.updateData(results)
+            } else {
+                adapter.addItems(results)
+            }
         } else {
-            progressLayout.showContent()
+            adapter.loadMoreListener = null
+            adapter.notifyDataSetChanged()
+        }
+
+        if (adapter.items.isEmpty()) {
+            if (isFragmentActive() && progressLayout != null)
+                progressLayout.showEmpty(
+                    resId = R.raw.empty_search,
+                    isLottieAnimation = true,
+                    emptyText = "No Novels Found!",
+                    buttonText = getString(R.string.try_again),
+                    onClickListener = View.OnClickListener {
+                        progressLayout.showLoading()
+                        currentPageNumber = 1
+                        searchNovels()
+                    })
+        } else {
+            if (isFragmentActive() && progressLayout != null)
+                progressLayout.showContent()
+        }
+    }
+
+    override fun loadMore() {
+        if (isPageLoading.compareAndSet(false, true)) {
+            currentPageNumber++
+            searchNovels()
         }
     }
 
     //region Adapter Listener Methods - onItemClick(), viewBinder()
 
-    override fun onItemClick(item: Novel) {
-        activity?.startNovelDetailsActivity(item, false)
+    override fun onItemClick(item: Novel, position: Int) {
+        (activity as? AppCompatActivity)?.startNovelDetailsActivity(item, false)
         //addToDownloads(item)
     }
 
     override fun bind(item: Novel, itemView: View, position: Int) {
         itemView.novelImageView.setImageResource(android.R.color.transparent)
 
-        if (item.imageUrl != null) {
+        if (!item.imageUrl.isNullOrBlank()) {
             Glide.with(this)
-                    .load(item.imageUrl?.getGlideUrl())
-                    .apply(RequestOptions.circleCropTransform())
-                    .into(itemView.novelImageView)
+                .load(item.imageUrl?.getGlideUrl())
+                .apply(RequestOptions.circleCropTransform())
+                .into(itemView.novelImageView)
         }
 
         //Other Data Fields
         itemView.novelTitleTextView.text = item.name
+        itemView.novelTitleTextView.isSelected = dataCenter.enableScrollingText
+
         if (item.rating != null && item.rating != "N/A") {
             var ratingText = "(N/A)"
             try {
@@ -149,7 +201,7 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
                 itemView.novelRatingBar.rating = rating
                 ratingText = "(" + String.format("%.1f", rating) + ")"
             } catch (e: Exception) {
-                Log.w("Library Activity", "Rating: " + item.rating, e)
+                Logs.warning(TAG, "Rating: " + item.rating, e)
             }
             itemView.novelRatingText.text = ratingText
         }
@@ -157,14 +209,9 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
 
 //endregion
 
-    override fun onDestroy() {
-        super.onDestroy()
-        async.cancelAll()
-    }
-
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        if (adapter.items.isNotEmpty())
+        if (::adapter.isInitialized && adapter.items.isNotEmpty())
             outState.putSerializable("results", adapter.items)
     }
 
