@@ -1,32 +1,44 @@
 package io.github.gmathi.novellibrary.fragment
 
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import co.metalab.asyncawait.async
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.lifecycleScope
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import io.github.gmathi.novellibrary.R
-import io.github.gmathi.novellibrary.activity.startNovelDetailsActivity
 import io.github.gmathi.novellibrary.adapter.GenericAdapter
-import io.github.gmathi.novellibrary.model.Novel
+import io.github.gmathi.novellibrary.dataCenter
+import io.github.gmathi.novellibrary.extensions.*
+import io.github.gmathi.novellibrary.model.database.Novel
 import io.github.gmathi.novellibrary.network.NovelApi
 import io.github.gmathi.novellibrary.network.searchUrl
+import io.github.gmathi.novellibrary.util.Logs
 import io.github.gmathi.novellibrary.util.Utils
 import io.github.gmathi.novellibrary.util.getGlideUrl
 import io.github.gmathi.novellibrary.util.setDefaults
+import io.github.gmathi.novellibrary.util.system.isFragmentActive
+import io.github.gmathi.novellibrary.util.system.startNovelDetailsActivity
 import kotlinx.android.synthetic.main.content_recycler_view.*
 import kotlinx.android.synthetic.main.listitem_novel.view.*
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import java.util.concurrent.atomic.AtomicBoolean
 
 
-class SearchUrlFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
+class SearchUrlFragment : BaseFragment(), GenericAdapter.Listener<Novel>, GenericAdapter.LoadMoreListener {
 
+    override var currentPageNumber: Int = 1
+    override val preloadCount: Int = 50
+    override val isPageLoading: AtomicBoolean = AtomicBoolean(false)
     private lateinit var searchUrl: String
 
     companion object {
+        private const val TAG = "SearchUrlFragment"
+
         fun newInstance(url: String): SearchUrlFragment {
             val bundle = Bundle()
             bundle.putString("url", url)
@@ -50,7 +62,7 @@ class SearchUrlFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         //(activity as AppCompatActivity).setSupportActionBar(null)
-        searchUrl = arguments!!.getString("url")
+        searchUrl = arguments?.getString("url")!!
         setRecyclerView()
 
         if (savedInstanceState != null) {
@@ -66,33 +78,36 @@ class SearchUrlFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
     }
 
     private fun setRecyclerView() {
-        adapter = GenericAdapter(items = ArrayList(), layoutResId = R.layout.listitem_novel, listener = this)
+        adapter = GenericAdapter(items = ArrayList(), layoutResId = R.layout.listitem_novel, listener = this, loadMoreListener = this)
         recyclerView.setDefaults(adapter)
         swipeRefreshLayout.setOnRefreshListener { searchNovels() }
     }
 
     private fun searchNovels() {
 
-        async search@ {
+        lifecycleScope.launch search@{
 
             if (!Utils.isConnectedToNetwork(activity)) {
-                progressLayout.showError(ContextCompat.getDrawable(context!!, R.drawable.ic_warning_white_vector), getString(R.string.no_internet), getString(R.string.try_again), {
+                progressLayout.noInternetError(View.OnClickListener {
                     progressLayout.showLoading()
+                    currentPageNumber = 1
                     searchNovels()
                 })
                 return@search
             }
 
-            val results = await { NovelApi.searchUrl(searchUrl) }
+            val results = withContext(Dispatchers.IO) { NovelApi.searchUrl(searchUrl, pageNumber = currentPageNumber) }
             if (results != null) {
                 if (isVisible && (!isDetached || !isRemoving)) {
                     loadSearchResults(results)
+                    isPageLoading.lazySet(false)
                     swipeRefreshLayout.isRefreshing = false
                 }
             } else {
                 if (isFragmentActive() && progressLayout != null)
-                    progressLayout.showError(ContextCompat.getDrawable(context!!, R.drawable.ic_warning_white_vector), getString(R.string.no_internet), getString(R.string.try_again), {
+                    progressLayout.showError(errorText = getString(R.string.connection_error), buttonText = getString(R.string.try_again), onClickListener = View.OnClickListener {
                         progressLayout.showLoading()
+                        currentPageNumber = 1
                         searchNovels()
                     })
             }
@@ -100,13 +115,30 @@ class SearchUrlFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
     }
 
     private fun loadSearchResults(results: ArrayList<Novel>) {
-        adapter.updateData(results)
+
+        if (results.isNotEmpty() && !adapter.items.containsAll(results)) {
+            if (currentPageNumber == 1) {
+                adapter.updateData(results)
+            } else {
+                adapter.addItems(results)
+            }
+        } else {
+            adapter.loadMoreListener = null
+            adapter.notifyDataSetChanged()
+        }
+
         if (adapter.items.isEmpty()) {
             if (isFragmentActive() && progressLayout != null)
-                progressLayout.showError(ContextCompat.getDrawable(context!!, R.drawable.ic_youtube_searched_for_white_vector), "No Novels Found!", "Try Again", {
-                    progressLayout.showLoading()
-                    searchNovels()
-                })
+                progressLayout.showEmpty(
+                    resId = R.raw.empty_search,
+                    isLottieAnimation = true,
+                    emptyText = "No Novels Found!",
+                    buttonText = getString(R.string.try_again),
+                    onClickListener = View.OnClickListener {
+                        progressLayout.showLoading()
+                        currentPageNumber = 1
+                        searchNovels()
+                    })
         } else {
             if (isFragmentActive() && progressLayout != null)
                 progressLayout.showContent()
@@ -115,22 +147,24 @@ class SearchUrlFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
 
 //region Adapter Listener Methods - onItemClick(), viewBinder()
 
-    override fun onItemClick(item: Novel) {
-        activity?.startNovelDetailsActivity(item, false)
+    override fun onItemClick(item: Novel, position: Int) {
+        (activity as? AppCompatActivity)?.startNovelDetailsActivity(item, false)
         //addToDownloads(item)
     }
 
     override fun bind(item: Novel, itemView: View, position: Int) {
         itemView.novelImageView.setImageResource(android.R.color.transparent)
-        if (item.imageUrl != null) {
+        if (!item.imageUrl.isNullOrBlank()) {
             Glide.with(this)
-                    .load(item.imageUrl?.getGlideUrl())
-                    .apply(RequestOptions.circleCropTransform())
-                    .into(itemView.novelImageView)
+                .load(item.imageUrl?.getGlideUrl())
+                .apply(RequestOptions.circleCropTransform())
+                .into(itemView.novelImageView)
         }
 
         //Other Data Fields
         itemView.novelTitleTextView.text = item.name
+        itemView.novelTitleTextView.isSelected = dataCenter.enableScrollingText
+
         if (item.rating != null) {
             var ratingText = "(N/A)"
             try {
@@ -138,22 +172,24 @@ class SearchUrlFragment : BaseFragment(), GenericAdapter.Listener<Novel> {
                 itemView.novelRatingBar.rating = rating
                 ratingText = "(" + String.format("%.1f", rating) + ")"
             } catch (e: Exception) {
-                Log.w("Library Activity", "Rating: " + item.rating, e)
+                Logs.warning(TAG, "Rating: " + item.rating, e)
             }
             itemView.novelRatingText.text = ratingText
         }
     }
 
-//endregion
-
-    override fun onDestroy() {
-        super.onDestroy()
-        async.cancelAll()
+    override fun loadMore() {
+        if (isPageLoading.compareAndSet(false, true)) {
+            currentPageNumber++
+            searchNovels()
+        }
     }
+
+//endregion
 
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        if (adapter.items.isNotEmpty())
+        if (::adapter.isInitialized && adapter.items.isNotEmpty())
             outState.putSerializable("results", adapter.items)
     }
 
