@@ -1,21 +1,21 @@
 package io.github.gmathi.novellibrary.network
 
-import io.github.gmathi.novellibrary.model.RecenlytUpdatedItem
+import io.github.gmathi.novellibrary.model.other.RecentlyUpdatedItem
 
 
-fun NovelApi.getRecentlyUpdatedNovels(): ArrayList<RecenlytUpdatedItem>? {
-    var searchResults: ArrayList<RecenlytUpdatedItem>? = null
+fun NovelApi.getRecentlyUpdatedNovels(): ArrayList<RecentlyUpdatedItem>? {
+    var searchResults: ArrayList<RecentlyUpdatedItem>? = null
     try {
         searchResults = ArrayList()
-        val document = getDocumentWithUserAgent("http://www.novelupdates.com/")
+        val document = getDocument("https://www.novelupdates.com/")
         val elements = document.body()?.getElementsByTag("td")?.filter { it.className().contains("sid") }
         if (elements != null)
             for (element in elements) {
-                val item = RecenlytUpdatedItem()
-                item.novelUrl = element.getElementsByTag("a")?.firstOrNull()?.attr("href")
-                item.novelName = element.getElementsByTag("a")?.firstOrNull()?.attr("title")
-                item.chapterName = element.getElementsByTag("a")?.firstOrNull { it.className() == "chp-release" }?.text()
-                item.publisherName = element.getElementsByTag("span")?.firstOrNull { it.className() == "mob_group" }?.getElementsByTag("a")?.firstOrNull()?.attr("title")
+                val item = RecentlyUpdatedItem()
+                item.novelUrl = element.selectFirst("a[href]")?.attr("abs:href")
+                item.novelName = element.selectFirst("a[title]")?.attr("title")
+                item.chapterName = element.selectFirst("a.chp-release")?.text()
+                item.publisherName = element.selectFirst("span.mob_group > a")?.attr("title")
 
                 searchResults.add(item)
             }
